+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidTransactionProfiler.java
@@ -28,7 +28,6 @@ import java.util.UUID;
 import java.util.concurrent.Future;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.jetbrains.annotations.TestOnly;
 
 final class AndroidTransactionProfiler implements ITransactionProfiler {
 
@@ -304,7 +303,10 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
         options.getProguardUuid(),
         versionName,
         versionCode,
-        options.getEnvironment());
+        options.getEnvironment(),
+        isTimeout
+            ? ProfilingTraceData.TRUNCATION_REASON_TIMEOUT
+            : ProfilingTraceData.TRUNCATION_REASON_NORMAL);
   }
 
   /**
@@ -326,10 +328,5 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       options.getLogger().log(SentryLevel.ERROR, "Error getting MemoryInfo.", e);
       return null;
     }
-  }
-
-  @TestOnly
-  void setTimedOutProfilingData(@Nullable ProfilingTraceData data) {
-    this.timedOutProfilingData = data;
   }
 }