+++ sentry-android-core/src/test/java/io/sentry/android/core/AndroidTransactionProfilerTest.kt
@@ -4,12 +4,17 @@ import android.content.Context
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
+import com.nhaarman.mockitokotlin2.spy
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.ILogger
+import io.sentry.ISentryExecutorService
+import io.sentry.ProfilingTraceData
 import io.sentry.SentryLevel
 import io.sentry.SentryTracer
 import io.sentry.TransactionContext
@@ -39,7 +44,7 @@ class AndroidTransactionProfilerTest {
             whenever(it.sdkInfoVersion).thenReturn(Build.VERSION_CODES.LOLLIPOP)
         }
         val mockLogger = mock<ILogger>()
-        val options = SentryAndroidOptions().apply {
+        val options = spy(SentryAndroidOptions()).apply {
             dsn = mockDsn
             profilesSampleRate = 1.0
             isDebug = true
@@ -229,23 +234,47 @@ class AndroidTransactionProfilerTest {
     fun `onTransactionFinish returns timedOutData to the timed out transaction once, even after other transactions`() {
         val profiler = fixture.getSut(context)
 
+        val executorService = mock<ISentryExecutorService>()
+        val captor = argumentCaptor<Runnable>()
+        whenever(executorService.schedule(captor.capture(), any())).thenReturn(null)
+        whenever(fixture.options.executorService).thenReturn(executorService)
         // Start and finish first transaction profiling
         profiler.onTransactionStart(fixture.transaction1)
-        val traceData = profiler.onTransactionFinish(fixture.transaction1)
 
-        // Set timed out data
-        profiler.setTimedOutProfilingData(traceData)
+        // Set timed out data by calling the timeout scheduled job
+        captor.firstValue.run()
 
-        // Start and finish second transaction profiling
+        // Start and finish second transaction. Since profiler returned data, it means no other profiling is running
         profiler.onTransactionStart(fixture.transaction2)
         assertEquals(fixture.transaction2.eventId.toString(), profiler.onTransactionFinish(fixture.transaction2)!!.transactionId)
 
         // First transaction finishes: timed out data is returned
-        val traceData2 = profiler.onTransactionFinish(fixture.transaction1)
-        assertEquals(traceData, traceData2)
+        val traceData = profiler.onTransactionFinish(fixture.transaction1)
+        assertEquals(traceData!!.transactionId, fixture.transaction1.eventId.toString())
 
         // If first transaction is finished again, nothing is returned
         assertNull(profiler.onTransactionFinish(fixture.transaction1))
+    }
+
+    @Test
+    fun `timedOutData has timeout truncation reason`() {
+        val profiler = fixture.getSut(context)
+
+        val executorService = mock<ISentryExecutorService>()
+        val captor = argumentCaptor<Runnable>()
+        whenever(executorService.schedule(captor.capture(), any())).thenReturn(null)
+        whenever(fixture.options.executorService).thenReturn(executorService)
+
+        // Start and finish first transaction profiling
+        profiler.onTransactionStart(fixture.transaction1)
+
+        // Set timed out data by calling the timeout scheduled job
+        captor.firstValue.run()
+
+        // First transaction finishes: timed out data is returned
+        val traceData = profiler.onTransactionFinish(fixture.transaction1)
+        assertEquals(traceData!!.transactionId, fixture.transaction1.eventId.toString())
+        assertEquals(ProfilingTraceData.TRUNCATION_REASON_TIMEOUT, traceData.truncationReason)
     }
 
     @Test